fix(text): carry run font size as fractional points
TextRun.SizePt was an int, so the Python bridge could not send a
non-integer point size: font.size = <EMU> converts to points and failed
with "cannot unmarshal number 0.0018897 into Go struct field
TextRun.paragraphs.runs.size_pt of type int". Reading was lossy too --
parse_text divided the sz attribute by 100 with integer division, so a
half-point size in an existing deck came back as a whole point.

SizePt and TextRunSpec.SizePt are float64 now, and sz is rounded to the
nearest centipoint on write, so 11.5pt round-trips. Boundaries that stay
whole points (text.Run, SlideContent.TitleSize) round explicitly.

list_placeholders also reports each placeholder's text, so callers can
read placeholder content without a separate shape lookup.

+++ internal/pptxxml/fuzz_test.go
@@ -186,7 +186,7 @@ func FuzzRichTextRun(f *testing.F) {
 			Color:         color,
 			Highlight:     highlight,
 			Font:          font,
-			SizePt:        sizePt,
+			SizePt:        float64(sizePt),
 			Code:          code,
 			AllCaps:       allCaps,
 			SmallCaps:     smallCaps,
